+++ frontend/src/pages/Esi.vue
@@ -157,7 +157,7 @@ export default {
         charSearchDelayed: _.debounce((vm, searchTerm) => {
             vm.charSearchResult = [];
             vm.charSearchIsLoading = true;
-            new CharacterApi().findBy(searchTerm, function(error, data) {
+            new CharacterApi().findCharacter(searchTerm, function(error, data) {
                 vm.charSearchIsLoading = false;
                 if (error) {
                     return;