+++ frontend/src/pages/UserAdmin.vue
@@ -76,7 +76,7 @@
                 <h3 class="card-header">Characters</h3>
                 <div class="card-body">
                     <!--suppress HtmlUnknownTag -->
-                    <character-search v-on:result="onSearchResult($event)"></character-search>
+                    <character-search v-on:result="onSearchResult($event)" :admin="true"></character-search>
                     <span class="text-muted small">
                         Select a character to show it's player account.
                     </span>